**huge** enum/struct refactoring
