+++ src/encrypt.rs
@@ -1,11 +1,11 @@
 use crate::encrypt::crypto::encrypt_bytes_memory_mode;
 use crate::encrypt::crypto::encrypt_bytes_stream_mode;
 use crate::file::get_bytes;
-use crate::global::parameters::Algorithm;
-use crate::global::parameters::BenchMode;
-use crate::global::parameters::CryptoParams;
-use crate::global::parameters::EraseMode;
-use crate::global::parameters::OutputFile;
+use crate::global::enums::Algorithm;
+use crate::global::enums::BenchMode;
+use crate::global::structs::CryptoParams;
+use crate::global::enums::EraseMode;
+use crate::global::enums::OutputFile;
 use crate::global::BLOCK_SIZE;
 use crate::key::get_secret;
 use crate::prompt::overwrite_check;
