+++ src/global/parameters.rs
@@ -5,183 +5,14 @@
 
 use anyhow::{Context, Result};
 use clap::ArgMatches;
-use std::fs::File;
-use std::io::Write;
-
-use super::SALT_LEN;
-
-pub struct CryptoParams {
-    pub hash_mode: HashMode,
-    pub skip: SkipMode,
-    pub bench: BenchMode,
-    pub password: PasswordMode,
-    pub erase: EraseMode,
-    pub keyfile: KeyFile,
-}
-
-// the information needed to easily serialize a header
-pub struct HeaderType {
-    pub header_version: HeaderVersion,
-    pub cipher_mode: CipherMode,
-    pub algorithm: Algorithm,
-}
-
-pub enum HeaderVersion {
-    V1,
-}
-
-// the data used returned after reading/deserialising a header
-pub struct Header {
-    pub header_type: HeaderType,
-    pub nonce: Vec<u8>,
-    pub salt: [u8; SALT_LEN],
-}
-
-pub struct PackMode {
-    pub dir_mode: DirectoryMode,
-    pub hidden: HiddenFilesMode,
-    pub exclude: Vec<String>,
-    pub compression_level: i32,
-    pub print_mode: PrintMode,
-}
-
-#[derive(PartialEq, Eq, Clone, Copy)]
-pub enum DirectoryMode {
-    Singular,
-    Recursive,
-}
-
-#[derive(PartialEq)]
-pub enum HiddenFilesMode {
-    Include,
-    Exclude,
-}
-
-#[derive(PartialEq)]
-pub enum PrintMode {
-    Verbose,
-    Quiet,
-}
-
-#[derive(PartialEq, Clone, Copy)]
-pub enum EraseMode {
-    EraseFile(i32),
-    IgnoreFile(i32),
-}
-
-#[derive(PartialEq, Clone, Copy)]
-pub enum HashMode {
-    CalculateHash,
-    NoHash,
-}
-
-#[derive(PartialEq, Copy, Clone)]
-pub enum SkipMode {
-    ShowPrompts,
-    HidePrompts,
-}
-
-#[derive(PartialEq, Copy, Clone)]
-pub enum BenchMode {
-    WriteToFilesystem,
-    BenchmarkInMemory,
-}
-
-#[derive(PartialEq, Copy, Clone)]
-pub enum PasswordMode {
-    ForceUserProvidedPassword,
-    NormalKeySourcePriority,
-}
-
-pub enum OutputFile {
-    Some(File),
-    None,
-}
-
-#[derive(PartialEq)]
-pub enum KeyFile {
-    Some(String),
-    None,
-}
-
-pub enum HeaderFile {
-    Some(String),
-    None,
-}
-
-#[derive(Copy, Clone)]
-pub enum Algorithm {
-    Aes256Gcm,
-    XChaCha20Poly1305,
-    DeoxysII256,
-}
+use crate::global::enums::*;
+use crate::global::structs::*;
 
 pub const ALGORITHMS: [Algorithm; 3] = [
     Algorithm::XChaCha20Poly1305,
     Algorithm::Aes256Gcm,
     Algorithm::DeoxysII256,
 ];
-
-impl EraseMode {
-    pub fn get_passes(self) -> i32 {
-        match self {
-            EraseMode::EraseFile(passes) => passes,
-            EraseMode::IgnoreFile(_) => 0,
-        }
-    }
-}
-
-impl KeyFile {
-    pub fn get_contents(&self) -> Result<String> {
-        match self {
-            KeyFile::Some(data) => Ok(data.to_string()),
-            KeyFile::None => Err(anyhow::anyhow!(
-                "Tried using a keyfile when one wasn't provided"
-            )), // should never happen
-        }
-    }
-}
-
-impl OutputFile {
-    pub fn write_all(&mut self, buf: &[u8]) -> std::io::Result<()> {
-        match self {
-            OutputFile::Some(file) => file.write_all(buf),
-            OutputFile::None => Ok(()),
-        }
-    }
-    pub fn flush(&mut self) -> std::io::Result<()> {
-        match self {
-            OutputFile::Some(file) => file.flush(),
-            OutputFile::None => Ok(()),
-        }
-    }
-}
-
-impl std::fmt::Display for Algorithm {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match *self {
-            Algorithm::Aes256Gcm => write!(f, "AES-256-GCM"),
-            Algorithm::XChaCha20Poly1305 => write!(f, "XChaCha20-Poly1305"),
-            Algorithm::DeoxysII256 => write!(f, "Deoxys-II-256"),
-        }
-    }
-}
-
-#[derive(PartialEq, Eq)]
-pub enum CipherMode {
-    // could do with a better name
-    MemoryMode,
-    StreamMode,
-}
-
-impl std::fmt::Display for CipherMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match *self {
-            CipherMode::MemoryMode => write!(f, "memory mode"),
-            CipherMode::StreamMode => write!(f, "stream mode"),
-        }
-    }
-}
 
 pub fn get_param(name: &str, sub_matches: &ArgMatches) -> Result<String> {
     let value = sub_matches
