+++ src/main.rs
@@ -1,7 +1,8 @@
 use anyhow::Result;
-use global::parameters::{get_param, SkipMode};
+use global::parameters::get_param;
 use list::show_values;
 use std::result::Result::Ok;
+use global::enums::SkipMode;
 
 mod cli;
 mod decrypt;
